requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdfplumber
python-multipart
httpx
